CanDeactivate Route Guard in Angular
The Angular CanDeactivate route guard is called, whenever we navigate away from the route before the current component gets deactivated.

The best use case for CanDectivate route guard is the form component. The user may have filled the data in the form and tries to leave that component without saving his work. The CanDeactivate guard gives us a chance to warn the user that he has not saved his work and give him a chance to cancel the navigation.